requirements.txt
blinker==1.4
Cerberus==1.3.4
click==8.1.2
Flask==2.1.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
PyJWT==2.3.0
pyparsing==3.0.8
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
